# [Model][Speculative Decoding] Integrate PARD into vLLM