## Description:

This PR integrates PARD into vLLM. PARD (PARallel Draft model) is a speculative decoding method that enables low-cost adaptation of autoregressive draft models into parallel draft models. It improves inference efficiency by allowing the draft model to predict multiple future tokens in a single forward pass, significantly reducing decoding latency. For detailed technical information, please refer to the [technical report](https://arxiv.org/pdf/2504.18583) and [blog](https://www.amd.com/en/developer/resources/technical-articles/accelerating-generative-llms-interface-with-parallel-draft-model-pard.html).


<p align="center">
  <figure style="display: inline-block; text-align: center;">
    <img src="https://cdn-uploads.huggingface.co/production/uploads/630cb01cc169245d78fe76b6/Dh-7wE-l0YAfU9lXWssKf.png" width="100%">
    <figcaption style="font-style: italic; margin-top: 2px;">
      AR and AR+ represent baseline auto-regressive generation using Transformers and Transformers+, respectively. VSD denotes vanilla speculative decoding. PARD refers to the proposed method in this work.
    </figcaption>
  </figure>
</p>

## Support Model Series
Supports acceleration for models across various sizes in the following series: Llama3, Deepseek-R1-distilled-Qwen, and Qwen1.5/2/2.5.
| Model Series | Draft Model Name                            | link      |
|--------------|---------------------------------------|---------------|
| llama3       | PARD-Llama-3.2-1B                     | [🤗 HuggingFace](https://huggingface.co/amd/PARD-Llama-3.2-1B)  |
| DSR Qwen     | PARD-DeepSeek-R1-Distill-Qwen-1.5B    | [🤗 HuggingFace](https://huggingface.co/amd/PARD-DeepSeek-R1-Distill-Qwen-1.5B) |
| Qwen         | PARD-Qwen2.5-0.5B                     | [🤗 HuggingFace](https://huggingface.co/amd/PARD-Qwen2.5-0.5B) |

## Summary of changes:
### 1. **vllm/spec_decode/multi_step_worker.py**:
 Added the **pard_infer** function to support PARD-based speculative decoding. Key logic includes:
- **KV cache recomputation & seq_group_metadata_list update**: Some of PARD's KV cache is derived from mask_token and needs to be recomputed. New mask tokens are introduced accordingly.

- **Proposal generation**: Uses the SpeculativeProposals function to obtain token proposals.

- **Draft model forward**: Only a single forward pass is needed for the draft model to generate multiple tokens.

- **Logits shape alignment**: Aligns the logits shapes of the target and draft models when they differ.

- **Output conversion**: Converts intermediate results into the standard output format.

### 2. **vllm/spec_decode/batch_expansion.py**:
Modified score_proposal to support:

- **keep_index**: If not None, only returns results for specified indices. Recomputed KV cache outputs will be excluded.

- **return_output**: If True, returns the full target_sampler_output.

### 3. **vllm/spec_decode/spec_decode_worker.py**:
Added a new interface to enable PARD inference.

## Test
### Test Code
```
import argparse
import json
import os
from vllm import LLM, SamplingParams
import requests
import numpy as np

os.environ.update({
    "VLLM_USE_V1": "0"
})
def parse_args():
    parser = argparse.ArgumentParser()
    parser.add_argument("--model", type=str, default="unsloth/Meta-Llama-3.1-8B-Instruct")
    parser.add_argument("--draft", type=str, default="amd/PARD-Llama-3.2-1B")
    parser.add_argument("--benchmark", type=str, default="humaneval")
    parser.add_argument("--max_num_seqs", type=int, default=1)
    parser.add_argument("--num_prompts", type=int, default=80)
    parser.add_argument("--num_spec_tokens", type=int, default=16)
    parser.add_argument("--tp", type=int, default=1)
    parser.add_argument("--temp", type=float, default=0)
    parser.add_argument("--ar", action='store_true')
    parser.add_argument("--disable-warmup", action='store_true')
    return parser.parse_args()

def main():
    args = parse_args()
    prompts = []
    for line in requests.get(f'https://raw.githubusercontent.com/AMD-AIG-AIMA/PARD/master/datas/bmk/{args.benchmark}.jsonl').text.splitlines():
        if line:
            prompts.append(json.loads(line)['data'])
    prompts = prompts[:args.num_prompts]
    datas = [[{"role": "system", "content": "You are a helpful assistant."}, {"role": "user", "content": prompt}] for prompt in prompts]
    llm = LLM(
        model=args.model,
        enable_prefix_caching=False,
        tensor_parallel_size=args.tp,
        max_num_batched_tokens=8192,
        max_model_len=8192,
        max_num_seqs=args.max_num_seqs,
        gpu_memory_utilization=0.8,
        speculative_config=None if args.ar else {
            "model": args.draft,
            "num_speculative_tokens": args.num_spec_tokens
            },
    )

    sampling_params = SamplingParams(temperature=args.temp, max_tokens=512)

    ## warmup
    if not args.disable_warmup:
        print("warmup...")
        outputs = llm.chat(
            datas,
            sampling_params=sampling_params,
        )

    # inference
    print("inference...")
    outputs = llm.chat(
        datas,
        sampling_params=sampling_params,
    )

    # speed
    speed = []
    for output in outputs:
        speed.append([len(output.outputs[0].token_ids), (output.metrics.finished_time - output.metrics.first_token_time)])
        print(f"[anwer]:\n {output.outputs[0].text}")

    print(f"\n\n{'='*100}\n\n")
    print(f'[speed]: {np.array(speed)[:,0].sum() / np.array(speed)[:,1].sum()}\n')

    # accepted
    if not args.ar:
        acceptance_counts = [0] * (args.num_spec_tokens + 1)
        for output in outputs:
            for step, count in enumerate(
                    output.metrics.spec_token_acceptance_counts):
                acceptance_counts[step] += count

        print(f"[acceptance length]: {1 + (sum(acceptance_counts) / acceptance_counts[0])}")
    print(f"\n\n{'='*100}\n\n")


if __name__ == "__main__":
    main()
```
### Test Result
- Target model: Meta-Llama-3.1-8B-Instruct
- Task: humaneval

| Device       |  Accept Length | Baseline AR t/s    | PARD t/s |
|-------------|---------|--------|-----------|
| MI300  | 7.2 | 181.29 | 416.30 |
| MI250    | 7.2 | 67.99 | 169.08 |
| H20    | 7.2 | 153.16 |   295.85 |