@aarnphm 2025-06-12
Hi @zihaoanllm, thanks for making this PR.

The code path you have here is v0. as of 0.9.x, all v0 codepath should be frozen and will only be merged for bugfix only. I would suggest implementing this under `vllm/v1`, as v0 code will be removed in 0.10.x and above