@zihaoanllm 2025-06-12
> Hi @zihaoanllm, thanks for making this PR.
> 
> The code path you have here is v0. as of 0.9.x, all v0 codepath should be frozen and will only be merged for bugfix only. I would suggest implementing this under `vllm/v1`, as v0 code will be removed in 0.10.x and above

Hi @aarnphm , thanks for your response.

Currently, we’ve implemented a basic version under v0. The speculative decoding feature and multi-model KV cache support in v1 are still under development. I do plan to integrate this into v1 in the future.

In the meantime, merging this method into v0 would be helpful for our current usage and would also facilitate the future integration into v1.