@zihaoanllm 2025-06-18
> Spec Decode in V1 should be supported, at least with eagle3 and medusa (limited support), but iiuc the multi-model KV cache is a requirements for this spec decode method?

While trying to integrate this into v1, I found that it currently seems to only support KV caches with a single shape [(code link)](https://github.com/vllm-project/vllm/blob/d4629dc43f92c189e90a3e2f2f8a52648aed4d9d/vllm/v1/core/kv_cache_utils.py#L946-L961)? Both PARD and vanilla SpecDec rely on using both a small draft model and large target model, and their attention structures are often different, which makes v1 incompatible for now. The eagle series works because the head structure matches the target model, and Medusa's draft model doesn't use a KV cache. 

I'm still getting familiar with vLLM, but I’d be happy to help with integrating this method into v1 in the future. Please feel free to correct me if I’ve misunderstood anything. Thanks!