@ekagra-ranjan 2025-06-19
@zihaoanllm - the result on [humaneval](https://github.com/vllm-project/vllm/pull/18541#issuecomment-2978578598) look good. Could you please also share AL on MTBench which will make it easier to [compare with EAGLE-1/3](https://github.com/vllm-project/vllm/pull/16937#issuecomment-2828889840)?